+++ RELEASE_NOTES.md
@@ -34,6 +34,7 @@ Apache Royale Compiler 0.9.11
 - compiler: Added CSS support for several translate, rotate, scale, skew, and matrix transformation functions in JS.
 - compiler: Add CSS support for declaring custom properties (CSS variables) and using `var` function in JS.
 - compiler: Fix crash when `[Style]` is of type `Object` and value is passed in MXML.
+- compiler: Fix null pointer exception when omitting quoted `<fx:Style>` value for `source` attribute.
 - debugger: Fix exception when evaluating certain expressions at run-time.
 - formatter: Added `insert-new-line-else` configuration option.
 - formatter: Filtered out unnecessary compiler warnings.